refactor(qbittorrent-e2e): rename module and split qbittorrent feature internals
