+++ src/console/ci/mod.rs
@@ -1,4 +1,4 @@
 //! Continuos integration scripts.
 pub mod compose;
 pub mod e2e;
-pub mod qbittorrent;
+pub mod qbittorrent_e2e;
