+++ src/console/ci/qbittorrent_e2e/client_role.rs
