+++ src/console/ci/qbittorrent_e2e/qbittorrent/credentials.rs
@@ -1,0 +1,8 @@
+/// Credentials for authenticating with the `qBittorrent` web UI.
+#[derive(Debug, Clone)]
+pub(crate) struct QbittorrentCredentials {
+    /// Web-UI username.
+    pub(crate) username: String,
+    /// Web-UI password.
+    pub(crate) password: String,
+}
