+++ src/console/ci/qbittorrent_e2e/scenario_steps/fixtures/build_payload_fixture.rs
