+++ src/console/ci/qbittorrent_e2e/scenario_steps/fixtures/build_torrent_fixture.rs
