+++ src/console/ci/qbittorrent_e2e/scenario_steps/qbittorrent/add_torrent_file_to_client.rs
@@ -1,6 +1,6 @@
 use anyhow::Context;
 
-use super::super::super::qbittorrent_client::QbittorrentClient;
+use super::super::super::qbittorrent::QbittorrentClient;
 
 /// Submits a `.torrent` file to a qBittorrent client.
 ///
