+++ src/console/ci/qbittorrent_e2e/scenario_steps/qbittorrent/wait_until_client_has_any_torrent.rs
@@ -1,5 +1,5 @@
 use super::super::super::poller::Poller;
-use super::super::super::qbittorrent_client::QbittorrentClient;
+use super::super::super::qbittorrent::QbittorrentClient;
 use super::super::super::types::{Deadline, PollInterval};
 
 /// Waits until the client reports at least one torrent in its list.
