+++ src/console/ci/qbittorrent_e2e/scenarios/mod.rs
