+++ src/console/ci/qbittorrent_e2e/types/compose_project_name.rs
