+++ src/console/ci/qbittorrent_e2e/types/poll_interval.rs
