+++ src/console/ci/qbittorrent_e2e/types/qbittorrent_image.rs
