+++ src/console/ci/qbittorrent_e2e/types/tracker_image.rs
